src/requirements.txt
pymilvus[milvus_lite]  # vector database (includes milvus-lite for local mode)
pymilvus[model]        # vector database
neo4j           # graph database
neo4j_graphrag  # graph rag
python-dotenv   # env variables
lmstudio        # local llm client
openai          # openai client
pdfminer.six    # pdf
PyPDF2          # pdf
pymupdf4llm     # pdf
docling         # chunking
markdown-chunker # chunking
deepdiff        # content differ
groq            # groq client